+++ MSP-codes/src/motors.h
@@ -15,7 +15,7 @@ This file include the header for the management of the Servo motors used for the
 #define LINK_2 110
 
 void init_servo(void);
-uint16_t angle_2_duty(uint16_t angle);                  // inputs: angle --> outputs: duty cycle
+uint16_t angle_2_duty(float angle);                  // inputs: angle --> outputs: duty cycle
 void set_servo(uint16_t duty1, uint16_t duty2);         // inputs: both of the duty cycles
 void set_pen();
 void set_position(pos_t pos);                           // inputs: position
